app/backend/functions/get-rating-viz/requirements.txt
functions_framework==3.4.0
numpy==1.24.3
protobuf==4.24.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-events==0.10.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
pandas==2.0.3
flask==2.3.2
seaborn==0.12.2
matplotlib==3.7.2
